split sortbar into its own component and have a top-bar around
