+++ shared/fs/folder/index.stories.js
@@ -7,6 +7,7 @@ import Folder from '.'
 import * as Kb from '../../common-adapters'
 import {rowsProvider} from '../row/index.stories'
 import {commonProvider} from '../common/index.stories'
+import {topBarProvider} from '../top-bar/index.stories'
 import {footerProvider} from '../footer/index.stories'
 import {bannerProvider} from '../banner/index.stories'
 import {headerProvider} from '../header/index.stories'
@@ -14,6 +15,7 @@ import {headerProvider} from '../header/index.stories'
 const provider = Sb.createPropProviderWithCommon({
   ...rowsProvider,
   ...commonProvider,
+  ...topBarProvider,
   ...footerProvider,
   ...bannerProvider,
   ...headerProvider,
