+++ shared/fs/index.stories.js
@@ -8,6 +8,7 @@ import footerStories from './footer/index.stories'
 import headerStories from './header/index.stories'
 import navHeaderStories from './nav-header/index.stories'
 import oopsStories from './oops/index.stories'
+import topBarStories from './top-bar/index.stories'
 import rowStories from './row/index.stories'
 import sendAttachmentToChatStories from './send-attachment-to-chat/index.stories'
 import sendLinkToChatStories from './send-link-to-chat/index.stories'
@@ -23,6 +24,7 @@ export default () =>
     headerStories,
     navHeaderStories,
     oopsStories,
+    topBarStories,
     rowStories,
     sendAttachmentToChatStories,
     sendLinkToChatStories,
